implemented yfinance and alpha vantage for financial data
