+++ agents/__init__.py
@@ -1,4 +1,4 @@
-# This file is used to import the AgentType and AgentFactory classes
-from .agent_factory import AgentType, AgentFactory
+# This file is used to import the AgentType and AIAgent classes
+from .ai_agent import AgentType, AIAgent
 
-__all__ = ['AgentType', 'AgentFactory']
+__all__ = ['AgentType', 'AIAgent']
